Getting latest mockito repository

+++ src/test/java/org/mockito/internal/invocation/InvocationImplTest.java
@@ -20,6 +20,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static junit.framework.TestCase.*;
+
 @SuppressWarnings({"unchecked"})
 public class InvocationImplTest extends TestBase {
 
@@ -156,7 +158,7 @@ public class InvocationImplTest extends TestBase {
             argTypes(int.class, int.class).args(1, argument).toInvocation();
 
         //when
-        int secondArgument = invocationOnInterface.getArgument(1);
+        int secondArgument = (Integer) invocationOnInterface.getArgument(1);
 
         //then
         assertTrue(secondArgument == argument);
